feat: #50 코드 분리
